correcao do problema 1040 e aumento de cobertura

+++ src/main/java/br/edu/ifpi/capar/uri/online/judge/URI_1040.java
@@ -17,7 +17,7 @@ public class URI_1040 {
         double nota2 = Double.parseDouble(numeros[1]);
         double nota3 = Double.parseDouble(numeros[2]);
         double nota4 = Double.parseDouble(numeros[3]);
-        double media = (nota1 * 2 + nota2 * 3 + nota3 * 4 + nota4 * 1) / 10;
+        double media = (float)((nota1 * 2 + nota2 * 3 + nota3 * 4 + nota4 * 1) / 10.0);
 
         System.out.println(String.format("Media: %.1f", media));
         if (media >= 7.0) {
@@ -28,7 +28,7 @@ public class URI_1040 {
             System.out.println("Aluno em exame.");
             double notaExame = Double.parseDouble(br.readLine());
             System.out.println(String.format("Nota do exame: %.1f", notaExame));
-            media = (media + notaExame) / 2;
+            media = (float)((media + notaExame) / 2.0);
             if (media >= 5.0) {
                 System.out.println("Aluno aprovado.");
             } else {
